+++ src/main/java/org/delia/core/FactoryService.java
@@ -2,6 +2,7 @@ package org.delia.core;
 
 import org.delia.assoc.DatIdMap;
 import org.delia.db.QueryBuilderService;
+import org.delia.db.hls.HLSSimpleQueryService;
 import org.delia.db.schema.SchemaMigrator;
 import org.delia.dval.compare.DValueCompareService;
 import org.delia.error.ErrorTracker;
@@ -33,6 +34,7 @@ public interface FactoryService {
 	LogFactory getLogFactory(); //may return null if not set in DeliaBuilder
 	ErrorTracker getErrorTracker();
 	QueryBuilderService getQueryBuilderService();
+	HLSSimpleQueryService createSimpleQueryService(ZDBInterfaceFactory dbInterface, DTypeRegistry registry);
 	SchemaMigrator createSchemaMigrator(ZDBInterfaceFactory dbInterface, DTypeRegistry registry, VarEvaluator varEvaluator, DatIdMap datIdMap);
 	ScalarValueBuilder createScalarValueBuilder(DTypeRegistry registry);
 	int getNextGeneratedRuleId();
