+++ src/test/java/org/delia/bdd/AllBDDTests.java
@@ -38,7 +38,6 @@ public class AllBDDTests extends BDDBase {
 		runR400File("t0-field-optional.txt", 4);
 		runR400File("t0-struct.txt", 4);
 		runR400File("t0-struct-inheritance.txt", 6);
-		runR400File("t0-field-optional.txt", 4);
 		runR400File("t0-field-primarykey.txt", 6);
 		runR400File("t0-field-serial.txt", 7);
 	}
