+++ src/test/java/org/delia/db/hls/HLSQueryBuilderTests.java
@@ -7,16 +7,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.delia.compiler.ast.QueryExp;
-import org.delia.core.FactoryService;
 import org.delia.db.DBHelper;
-import org.delia.db.QueryBuilderService;
-import org.delia.db.QueryBuilderServiceImpl;
-import org.delia.db.QueryContext;
-import org.delia.db.QuerySpec;
-import org.delia.db.hls.manager.HLSManagerBase;
 import org.delia.db.hls.manager.HLSManagerResult;
-import org.delia.queryresponse.LetSpanEngine;
-import org.delia.runner.DoNothingVarEvaluator;
 import org.delia.type.DType;
 import org.delia.type.DTypeRegistry;
 import org.delia.type.DValue;
@@ -33,55 +25,6 @@ import org.junit.Test;
  */
 public class HLSQueryBuilderTests extends HLSTestBase {
 	
-	public static class HLSQueryService { // implements QueryBuilderService {
-		private QueryBuilderService innerSvc;
-		private HLSManagerBase hlsManager;
-		private FactoryService factorySvc;
-		
-		public HLSQueryService(FactoryService factorySvc, HLSManagerBase hlsManager) {
-			this.factorySvc = factorySvc;
-			this.innerSvc = new QueryBuilderServiceImpl(factorySvc);
-			this.hlsManager = hlsManager;
-		}
-
-//		public HLSManagerResult execEqQuery(String typeName, String fieldName, DValue targetValue, ZDBExecutor zexec) {
-//			QueryExp queryExp = innerSvc.createEqQuery(typeName, fieldName, targetValue);
-//			return execQuery(queryExp, zexec);
-//		}
-//
-//		public HLSManagerResult execPrimaryKeyQuery(String typeName, DValue keyValue, ZDBExecutor zexec) {
-//			QueryExp queryExp = innerSvc.createPrimaryKeyQuery(typeName, keyValue);
-//			return execQuery(queryExp, zexec);
-//		}
-//
-//		public HLSManagerResult execInQuery(String typeName, List<DValue> list, DType relType, ZDBExecutor zexec) {
-//			QueryExp queryExp = innerSvc.createInQuery(typeName, list, relType);
-//			return execQuery(queryExp, zexec);
-//		}
-//
-//		public HLSManagerResult execCountQuery(String typeName, ZDBExecutor zexec) {
-//			QueryExp queryExp = innerSvc.createCountQuery(typeName);
-//			return execQuery(queryExp, zexec);
-//		}
-//
-//		public HLSManagerResult execAllRowsQuery(String typeName, ZDBExecutor zexec) {
-//			QueryExp queryExp = innerSvc.createAllRowsQuery(typeName);
-//			return execQuery(queryExp, zexec);
-//		}
-		public HLSManagerResult execQuery(QueryExp queryExp, ZDBExecutor zexec) {
-			QuerySpec querySpec = innerSvc.buildSpec(queryExp, new DoNothingVarEvaluator());
-			
-			QueryContext qtx = new QueryContext();
-			qtx.letSpanEngine = new LetSpanEngine(factorySvc, hlsManager.getRegistry()); 
-			HLSManagerResult result = hlsManager.execute(querySpec, qtx, zexec);
-			return result;
-		}
-
-		public QueryBuilderService getQueryBuilderSvc() {
-			return innerSvc;
-		}
-	}
-
 	@Test
 	public void testAllRows() {
 		try(ZDBExecutor dbexecutor = createExecutor()) {
@@ -153,8 +96,8 @@ public class HLSQueryBuilderTests extends HLSTestBase {
 	
 	
 	//---
-	private HLSManagerBase hlsManager;
-	private HLSQueryService svc;
+	//private HLSManagerBase hlsManager;
+	private HLSSimpleQueryService svc;
 	private DTypeRegistry registry;
 	
 	@Before
@@ -164,9 +107,8 @@ public class HLSQueryBuilderTests extends HLSTestBase {
 		compileQueryToLetStatement("");
 
 		registry = session.getExecutionContext().registry;
-		hlsManager = new HLSManagerBase(delia.getFactoryService(), delia.getDBInterface(), registry, new DoNothingVarEvaluator());
-		svc = new HLSQueryService(delia.getFactoryService(), hlsManager);
-		hlsManager.setGenerateSQLforMemFlag(true);
+//		svc = new HLSSimpleQueryService(delia.getFactoryService(), delia.getDBInterface(), registry);
+		svc = delia.getFactoryService().createSimpleQueryService(delia.getDBInterface(), registry);
 	}
 	
 	private void chkSqlGen(HLSManagerResult hlsResult, String sqlExpected) {
